GRPC-265 FEAT Separate health module from the server runtime.
