+++ grpc-health/src/main/java/io/micronaut/grpc/health/GrpcServerHealthIndicator.java
@@ -13,23 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.grpc.server.health;
+package io.micronaut.grpc.health;
+
+import static io.micronaut.core.util.CollectionUtils.mapOf;
+import static java.util.Objects.nonNull;
+
+import java.util.Map;
 
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.async.publisher.AsyncSingleResultPublisher;
 import io.micronaut.core.util.StringUtils;
-import io.micronaut.grpc.server.GrpcEmbeddedServer;
-import io.micronaut.grpc.server.GrpcServerConfiguration;
 import io.micronaut.health.HealthStatus;
 import io.micronaut.management.endpoint.health.HealthEndpoint;
 import io.micronaut.management.health.indicator.HealthIndicator;
 import io.micronaut.management.health.indicator.HealthResult;
-import org.reactivestreams.Publisher;
-
+import io.micronaut.runtime.server.EmbeddedServer;
+import jakarta.inject.Named;
 import jakarta.inject.Singleton;
-import java.util.Map;
-
-import static io.micronaut.core.util.CollectionUtils.mapOf;
+import org.reactivestreams.Publisher;
 
 /**
  * A {@link HealthIndicator} for Grpc server.
@@ -38,20 +39,20 @@ import static io.micronaut.core.util.CollectionUtils.mapOf;
  * @since 2.1.0
  */
 @Singleton
-@Requires(property = GrpcServerConfiguration.HEALTH_ENABLED, value = StringUtils.TRUE, defaultValue = StringUtils.TRUE)
+@Requires(property = GrpcHealthFactory.HEALTH_ENABLED, value = StringUtils.TRUE, defaultValue = StringUtils.TRUE)
 @Requires(beans = HealthEndpoint.class)
-@Requires(beans = GrpcEmbeddedServer.class)
+@Requires(beans = EmbeddedServer.class)
 public class GrpcServerHealthIndicator implements HealthIndicator {
     private static final String ID = "grpc-server";
 
-    private final GrpcEmbeddedServer server;
+    private final EmbeddedServer server;
 
     /**
      * Default constructor.
      *
      * @param server The grpc embedded server
      */
-    public GrpcServerHealthIndicator(GrpcEmbeddedServer server) {
+    public GrpcServerHealthIndicator(@Named("grpc") EmbeddedServer server) {
         this.server = server;
     }
 
@@ -72,8 +73,11 @@ public class GrpcServerHealthIndicator implements HealthIndicator {
          * it throws an unexpected exception that breaks the /health endpoint
          */
         final String serverHost = server.getHost();
-        final int serverPort = server.getServerConfiguration().getServerPort(); // don't call the server.getPort() here!
-        final Map details = mapOf("host", serverHost, "port", serverPort);
+        Integer serverPort = null;
+        if (server.isRunning()) { // don't call the server.getPort() here without run check!
+            serverPort = server.getPort();
+        }
+        final Map<?, ?> details = mapOf("host", serverHost, "port", nonNull(serverPort) ? serverPort : "N/A");
 
         return HealthResult
                 .builder(ID, healthStatus)
